+++ lib/localization/app_localization.dart
@@ -3,7 +3,6 @@ import 'dart:ui';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
-import 'package:intl/date_symbol_data_local.dart';
 import '/helpers/helper.dart';
 import '/widgets/ny_language_switcher.dart';
 
@@ -101,10 +100,6 @@ class NyLocalization {
 
     if (_assetsDir == null) {
       return;
-    }
-
-    if (_locale != null) {
-      initializeDateFormatting(_locale!.languageCode, null);
     }
 
     if (_assetsDir == null) {